Refactor time component and remove fake updates
